show_images/requirements.txt
aliyun-python-sdk-core==2.13.35
aliyun-python-sdk-ecs==4.24.3
aliyun-python-sdk-kms==2.15.0
aliyun-python-sdk-rds==2.5.11
appdirs==1.4.4
asgiref==3.4.1
Beaker==1.11.0
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.6
chardet==4.0.0
crcmod==1.7
cryptography==3.4.7
distlib==0.3.2
Django==3.2.5
filelock==3.0.12
google-api-core==1.27.0
google-auth==1.30.0
google-cloud-dialogflow-cx==0.4.1
googleapis-common-protos==1.53.0
grpcio==1.37.1
idna==2.10
jmespath==0.10.0
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==2.0.1
mysqlclient==2.0.3
oss2==2.14.0
packaging==20.9
pipenv==2021.5.29
proto-plus==1.18.1
protobuf==3.17.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
rsa==4.7.2
semantic-version==2.8.5
setuptools-rust==0.12.1.dev36+g255e073
six==1.16.0
sqlparse==0.4.1
style==1.1.0
tk==0.1.0
toml==0.10.2
typing-extensions==3.10.0.0
update==0.0.1
urllib3==1.26.4
virtualenv==20.4.7
virtualenv-clone==0.5.4
